Implement throttling of requests pr session id
This is an imperfect but possibly sufficient method
of preventing the services from issuing more requests
than the allowed rate of 3600 pr hour.

It works as long as the same session id instance
is shared among all services.

It will not work if multiple threads or processes
issues request with separate session id instances.